# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 1.0.2
components:
    securitySchemes:
        bearerAuth:            # arbitrary name for the security scheme
          type: http
          scheme: bearer
          bearerFormat: JWT    # optional, arbitrary value for documentation purposes
...
paths:
...
  /users/{idUser}/bans/{idUserBanned}:
    #unbanUser
    delete:
      security:
        - bearerAuth: []
      tags: ["ban"]
      operationId: unbanUser
      summary: unban an existing user
      description: |
        unban an existing (and already banned) user given its id
      parameters:
      - name: idUser
        in: path
        description: id of the user who is unbanning the other user
        schema:
          type: integer
          example: 1
        required: true
      - name: idUserBanned
        in: path
        description: id of the user to unban
        schema:
          type: integer
          example: 1
        required: true
      responses:
        "200":
          description: user succesfully unbanned
        "401":
          description: you must be logged in, not banned
        "403":
          description: you can't sban an user for another user/you can't sban yourself, not sbanned
          content:
            application/json:
              schema:
                description: returns the idUser, the idUserToBan and the token
                type: object
                properties:
                  idUser:
                    description: id of the user who is sbanning
                    type: integer
                    example: 1
                  idUserToSban:
                    description: if of the user to sban
                    type: integer
                    example: 2
                  token:
                    description: the authentication token
                    type: integer
                    example: 1
        "404":
          description: user (sbanner/to sban) not found, not sbanned
          content:
            application/json:
              schema:
                description: returns the idUser and the idUserToBan
                type: object
                properties:
                  idUser:
                    description: id of the user who is sbanning
                    type: integer
                    example: 1
                  idUserToSban:
                    description: if of the user to sban
                    type: integer
                    example: 2
...